use sync objects

+++ blade/gfx/renderer.cc
@@ -77,7 +77,7 @@ namespace blade
             return program_handle { BLADE_NULL_HANDLE };
         }
 
-        void renderer::set_viewport(const framebuffer_handle framebuffer, u32 x, u32 y, struct width width, struct height height) const noexcept
+        void renderer::set_viewport(const framebuffer_handle framebuffer, f32 x, f32 y, struct width width, struct height height) const noexcept
         {
             if (_backend)
             {
